deps/v8/test/inspector/debugger/stepping-async-generator-expected.txt

Async generator stepping

Running test: testStepIntoInitialYield
Setting breakpoint on call to generator()
Calling callGenerator()
function callGenerator() {
  #return generator();
}

Stepping into the generator()
function callGenerator() {
  return generator();#
}
